move parsing to own go file
