+++ parse_args.go
@@ -1,0 +1,18 @@
+package main
+
+import (
+	"fmt"
+	"path/filepath"
+)
+
+func parse_args(args []string) (string, error) {
+	if len(args) < 2 {
+		return "", fmt.Errorf("please add a root directory as an argument")
+	}
+
+	root, err := filepath.Abs(args[1])
+	if err != nil {
+		return "", err
+	}
+	return root, nil
+}
